+++ src/main/java/ru/viktorpenkin/fibonacci/dto/NumDTO.java
@@ -9,17 +9,17 @@ import lombok.Setter;
 @Getter
 @Setter
 public class NumDTO {
-    private Integer value;
+    private Long value;
 
-    public NumDTO(Integer value) {
+    public NumDTO(Long value) {
         this.value = value;
     }
 
-    public Integer getValue() {
+    public Long getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Long value) {
         this.value = value;
     }
 }
